refactor: separated one component from another
