+++ src/ts/components/CartList/CartList.ts
@@ -58,8 +58,7 @@ export class CartList {
       ]
     });
 
-    console.log('props - ', props);
-
+    let numInList = 0;
     const productItem =
       props?.items.map((item) => {
         const items = create({
@@ -69,7 +68,7 @@ export class CartList {
             create({
               tagName: 'div',
               classNames: 'product-list__item__number',
-              children: '1'
+              children: `${(numInList += 1)}`
             }),
             create({
               tagName: 'div',
@@ -187,78 +186,10 @@ export class CartList {
         return items;
       }) || [];
 
-    const buyBtn = create({
-      tagName: 'button',
-      classNames: 'btn',
-      children: 'BUY NOW'
-    });
-
-    const total = create({
-      tagName: 'div',
-      classNames: 'total',
-      children: [
-        create({
-          tagName: 'h3',
-          classNames: 'total__title h3',
-          children: 'Summary'
-        }),
-        create({
-          tagName: 'div',
-          classNames: 'total__info',
-          children: [
-            create({
-              tagName: 'div',
-              classNames: 'total__info__count',
-              children: [
-                `Products: `,
-                create({
-                  tagName: 'i',
-                  children: '10'
-                })
-              ]
-            }),
-            create({
-              tagName: 'div',
-              classNames: 'total__info__num',
-              children: [
-                `Total: `,
-                create({
-                  tagName: 'i',
-                  children: '1000$'
-                })
-              ]
-            }),
-            create({
-              tagName: 'div',
-              children: `<input class="total__info__promo-code promo-input input" type="search" name="search-promo" id="" placeholder="Enter promo code">`
-            }),
-            create({
-              tagName: 'div',
-              classNames: 'total__info__promo-code-hint',
-              children: [
-                `Promo for test: `,
-                create({
-                  tagName: 'i',
-                  children: [`"SHIT", "SHIT2"`]
-                })
-              ]
-            }),
-            buyBtn
-          ]
-        })
-      ]
-    });
-
-    const productList = create({
-      tagName: 'div',
-      classNames: 'product-list',
-      children: [header, ...productItem]
-    });
-
     this.component = create({
       tagName: 'div',
-      classNames: 'cart__wrapper',
-      children: [productList, total],
+      classNames: 'product-list',
+      children: [header, ...productItem],
       parent: this.parent
     });
   };
